# docs/api/teamflow-openapi.yaml
openapi: 3.0.3
info:
  title: TeamFlow API
  version: 0.3.0
  description: >
    TeamFlow のコアAPI仕様 (Auth / Projects / Tasks / Members / Invitations / Comments / Labels)。
    認証は cookie ベース (sid) を前提とする。
servers:
  - url: https://api.teamflow.example.com
    description: Production (仮)
  - url: http://localhost:3000
    description: Local dev (Next.js)
paths:
...
  /api/projects:
    get:
      summary: 自分が所属するプロジェクト一覧
      tags: [Projects]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: プロジェクト一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: 新規プロジェクト作成
      tags: [Projects]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: 作成されたプロジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/projects/{projectId}/tasks:
    get:
      summary: プロジェクト内タスク一覧（カンバン用）
      tags: [Tasks]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
        # --- Filters ---
        - name: status
          in: query
          required: false
          description: >
            ステータスでフィルタ。カンマ区切りで複数指定可能（例: status=todo,in_progress）。
            推奨値: todo / in_progress / done。
            互換のため doing も受け付ける（doing は in_progress に正規化される想定）。
          schema:
            type: string
          style: form
          explode: false
        - name: assigneeId
          in: query
          required: false
          description: 担当者のユーザーIDで絞り込み（UUID）。
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          required: false
          description: >
            優先度でフィルタ。カンマ区切りで複数指定可能（例: priority=high,medium）。
            推奨値: low / medium / high。
          schema:
            type: string
          style: form
          explode: false
        - name: dueDateFrom
          in: query
          required: false
          description: 期限の開始日（YYYY-MM-DD形式）。この日以降のタスクを取得
          schema:
            type: string
            format: date
        - name: dueDateTo
          in: query
          required: false
          description: 期限の終了日（YYYY-MM-DD形式）。この日以前のタスクを取得
          schema:
            type: string
            format: date
        - name: q
          in: query
          required: false
          description: 検索クエリ（MVPはタイトルの部分一致を想定）
          schema:
            type: string
            minLength: 1
        # --- Sorting ---
        - name: sort
          in: query
          required: false
          description: >
            ソート順を指定。形式: sort=-priority,createdAt（- はDESC、無印はASC）。
            使用可能キー: sortOrder, createdAt, updatedAt, dueDate, priority。
            dueDate の null 値は最後に寄せる（ASC時は最後、DESC時は最初）。
            priority は辞書順ではなく業務順（high > medium > low）でソートされる。
          schema:
            type: string
            example: "-priority,createdAt"
        # --- Safety / Limit ---
        - name: limit
          in: query
          required: false
          description: 取得件数の上限。未指定時は200、最大200件まで取得可能
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 200
        - name: cursor
          in: query
          required: false
          description: >
            Cursor-based pagination 用のカーソル（opaque）。
            前回のレスポンスで返された page.nextCursor をそのまま次回リクエストの cursor に指定してください。
            cursor を使用する場合、sort パラメータは指定できません（v1 の制限）。
          schema:
            type: string
      responses:
        '200':
          description: タスク一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  page:
                    type: object
                    description: ページング情報
                    properties:
                      nextCursor:
                        type: string
                        nullable: true
                        description: >
                          次ページ取得用のカーソル。null の場合は末尾（次ページなし）を表します。
                          この値を次回リクエストの cursor パラメータに指定してください。
                      limit:
                        type: integer
                        description: 取得件数の上限（リクエストで指定された limit 値）
                    required: [nextCursor, limit]
                required: [tasks]
        '400':
          description: クエリパラメータのバリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: タスク作成
      tags: [Tasks]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 作成されたタスク
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: プロジェクトに対する権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 内部サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{taskId}:
    get:
      summary: タスク詳細取得
      tags: [Tasks]
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: タスク詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: タスクが存在しない or 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: タスク更新（タイトル/説明/ステータス/期限など）
      tags: [Tasks]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: 更新後のタスク
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  schemas:
...
    ValidationIssue:
      type: object
      description: バリデーションエラーの1件分（どの入力のどこが問題か）
      properties:
        location:
          type: string
          description: エラー発生箇所
          enum: [query, path, body]
        field:
          type: string
          description: >
            対象フィールド名（例: status, priority, sort, dueDateFrom）
        code:
          type: string
          description: >
            機械判定用のエラーコード。
            主なコード:
            - INVALID_ENUM: 無効な列挙値
            - INVALID_FORMAT: 形式不正（例: cursor の形式不正、日付形式不正）
            - INVALID_RANGE: 範囲外の値（例: limit が範囲外）
            - CONSTRAINT_VIOLATION: 制約違反（例: dueDateFrom > dueDateTo）
            - INCOMPATIBLE_WITH_CURSOR: cursor と sort の併用（v1 では不許可）
            - INVALID_SIGNATURE: cursor の署名不一致（改ざん疑い）
            - EXPIRED: cursor の有効期限切れ
            - QUERY_MISMATCH: cursor のクエリ条件不一致（フィルタ等が変更された）
          example: INVALID_ENUM
        message:
          type: string
          description: >
            人間向けメッセージ（修正すべき内容が分かる文言）
        rejectedValue:
          nullable: true
          description: >
            受け付けられなかった入力値（出せる場合のみ）
      required: [location, field, code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: >
            エラー種別コード（例: VALIDATION_ERROR）
        message:
          type: string
          description: 人間向けメッセージ
        details:
          type: object
          description: フィールド単位の詳細エラーなど
          properties:
            issues:
              type: array
              description: バリデーションエラーの詳細（複数件）
              items:
                $ref: '#/components/schemas/ValidationIssue'
      required: [error, message]

    # -------- User --------
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ownerId, name, status, createdAt, updatedAt]
    ProjectCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, doing, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        assigneeId:
          type: string
          format: uuid
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - sortOrder
        - createdAt
        - updatedAt
    TaskCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, doing, in_progress, done]
          default: todo
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        assigneeId:
          type: string
          format: uuid
        dueDate:
          type: string
          format: date-time
      required: [title]
    TaskUpdateRequest:
      type: object
      minProperties: 1
      description: PATCH /api/tasks/{taskId} のリクエストボディ。すべてのフィールドは任意（optional）。更新したいフィールドのみを指定する。少なくとも1つのフィールドは必須。
      properties:
        title:
          type: string
          description: タスクのタイトル。空文字または空白のみの場合は400エラー。
        description:
          type: string
          nullable: true
          description: タスクの説明。
        status:
          type: string
          enum: [todo, doing, in_progress, done]
          description: タスクのステータス。正式な値は todo / in_progress / done。"doing" が指定された場合は互換のため "in_progress" に正規化される。
        priority:
          type: string
          enum: [low, medium, high]
          description: タスクの優先度。
        assigneeId:
          type: string
          format: uuid
          nullable: true
          description: 担当者のユーザーID。
        dueDate:
          type: string
          format: date-time
          nullable: true
          description: 期限日時（RFC3339形式）。
...